fix: 重构后日志 filter 错误导致不记录日志
