+++ package.json
@@ -16,23 +16,23 @@
     "update:all": "pnpm up --latest && cd static && pnpm up --latest && cd ../src-tauri && cargo update"
   },
   "dependencies": {
-    "@tauri-apps/api": "^2.0.0-beta.7",
+    "@tauri-apps/api": "2.0.0-beta.9",
     "@tauri-apps/plugin-clipboard-manager": "2.1.0-beta.0",
     "@tauri-apps/plugin-dialog": "2.0.0-beta.2",
     "@tauri-apps/plugin-shell": "2.0.0-beta.2",
     "@tauri-apps/plugin-updater": "2.0.0-beta.2",
     "alley-components": "^0.1.4",
     "solid-icons": "^1.1.0",
-    "solid-js": "^1.8.16"
+    "solid-js": "^1.8.17"
   },
   "devDependencies": {
-    "@types/node": "*",
     "@tauri-apps/cli": "2.0.0-beta.12",
+    "@types/node": "^20.12.7",
     "cross-env": "^7.0.3",
+    "internal-ip": "^8.0.0",
     "sass": "^1.75.0",
     "typescript": "^5.4.5",
-    "vite": "^5.2.8",
-    "vite-plugin-solid": "^2.10.2",
-    "internal-ip": "^7.0.0"
+    "vite": "^5.2.10",
+    "vite-plugin-solid": "^2.10.2"
   }
 }
